# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FPT Event Management API",
    "version": "1.0.0",
    "description": "API for FPT Event Management System - Go Lambda"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/api/registrations/my-tickets": {
      "get": {
        "tags": [
          "Tickets"
        ],
        "summary": "[Student] View registered ticket history",
        "description": "Get ticket list of logged in user with event and venue info",
        "operationId": "getMyTickets",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MyTicketResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not logged in"
          }
        }
      }
    },
...
    "/api/payment/my-bills": {
      "get": {
        "tags": [
          "Payment"
        ],
        "summary": "[Student] View payment history (Bills)",
        "description": "Get bill list with user name",
        "operationId": "getMyBills",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Bill list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Bill"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not logged in"
          }
        }
      }
    },
...
    "/api/buyTicket": {
      "get": {
        "tags": [
          "Payment"
        ],
        "summary": "VNPay return URL - confirm payment and create ticket",
        "description": "VNPay redirects here after payment. System verifies signature, checks response code, creates Bill and Ticket.",
        "operationId": "buyTicket",
        "parameters": [
          {
            "name": "vnp_Amount",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Amount from VNPay (x100)"
          },
          {
            "name": "vnp_ResponseCode",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Transaction result code (00 = success)"
          },
          {
            "name": "vnp_OrderInfo",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Encoded order info"
          },
          {
            "name": "vnp_TxnRef",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Transaction reference"
          },
          {
            "name": "vnp_SecureHash",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "HMAC SHA512 signature"
          }
        ],
        "responses": {
          "200": {
            "description": "Payment processing result",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "MyTicketResponse": {
        "type": "object",
        "properties": {
          "ticketId": {
            "type": "integer"
          },
          "ticketCode": {
            "type": "string"
          },
          "eventName": {
            "type": "string"
          },
          "venueName": {
            "type": "string"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "BOOKED",
              "CHECKED_IN",
              "CHECKED_OUT",
              "REFUNDED"
            ]
          },
          "checkInTime": {
            "type": "string",
            "format": "date-time"
          },
          "checkOutTime": {
            "type": "string",
            "format": "date-time"
          },
          "category": {
            "type": "string",
            "description": "Seat type"
          },
          "categoryPrice": {
            "type": "number"
          },
          "seatCode": {
            "type": "string"
          },
          "buyerName": {
            "type": "string"
          },
          "purchaseDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Bill": {
        "type": "object",
        "properties": {
          "billId": {
            "type": "integer"
          },
          "totalAmount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "paymentMethod": {
            "type": "string"
          },
          "paymentStatus": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "userName": {
            "type": "string"
          }
        }
      },
...
    }
  }
}